The model, AND skybox now load properly!
Inside the drawModel for loop, I was enabling the vertexattribarray before binding the VAO. This causes problems because I am disabling it at the end of the drawModel. So I disable, then I enable before binding the VAO. This causes problems because I cannot enable the pointer before binding the VAO because the pointer doesn't know what to point to. I think. So I either stop disabling, or I bind the VAO before enabling. I bound the VAO before enabling and I still disable at the end of the for loop